Update cdo-council-membership-status.csv
From DOC's /data page: "The Commerce Data Governance Board (CDGB) was established and first met in September 2019. The CDGB is the Department’s data governance and coordination body, and is chaired by the Commerce Chief Data Officer, with representatives from all Bureaus. The CDGB Terms of Reference were approved by its members in December 2019, and *the body meets monthly to coordinate the management of Commerce data as an asset*."